ramips: ignore already handled ethernet phys
The whole logic in fe_phy_connect() is based on the assumption that
mdio address and switch port id are equal. Albeit it is true for most
boards, it isn't is for all.

It isn't yet clear which subtargets/boards require the devicetree less
ethernet phy handling. Hence change the code in a way that it doesn't
touch ethernet phys which were early attached and handled.

Signed-off-by: Mathias Kresin <dev@kresin.me>