ramips: handle mdio address and switch port seperate
The phy handling code forces the phy mdio address and the switch port
(to which a phy is attached) to be the same. Albeit such a
configuration is used for most boards, it isn't for all.

Pass the switch port number to the ethernet phy connect functions, to
ensure the correct list entry is edited and not the list entry that
matches th phys mdio address.

Use the mdio address with mdiobus_get_phy instead of the port number,
to make sure the expected ethernet phy gets connected.

Signed-off-by: Mathias Kresin <dev@kresin.me>